requirements.txt
# ===============================
# 🧠 AI Financial Document Analyzer
# Backend: FastAPI  |  Frontend: Streamlit
# Agents: OCR + Reasoning + Gemini 2.5 Flash
# ===============================

# --- Backend Core ---
fastapi==0.121.0
uvicorn==0.38.0
python-dotenv==1.2.1
pydantic==2.12.4
starlette==0.49.3

# --- Data & Document Processing ---
pandas==2.3.3
pillow==12.0.0
pytesseract==0.3.13
PyMuPDF==1.26.6
pdf2image==1.17.0
psutil==7.1.3

# --- Google Gemini SDK ---
google-genai==1.49.0

# --- Utility & Core Networking ---
requests==2.32.5
protobuf==5.29.5
python-multipart==0.0.20
tenacity==9.1.2

# --- Frontend (Streamlit Visualization) ---
streamlit==1.51.0
altair==5.5.0
matplotlib==3.10.7

# --- Optional Enhancements (safe to keep) ---
annotated-types==0.7.0
typing-extensions==4.15.0
